+++ src/main/java/br/com/pegasus/api/rest/commerce/infra/repository/ProductRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
   Optional<ProductEntity> findByName(String name);
 }
